hello_docker/requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
tzdata==2022.1

djangorestframework~=3.13.1